requirements.txt
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0
soccerdata>=1.4.0
tqdm>=4.65.0

# Airflow providers (Airflow core đã có sẵn trong base image)
apache-airflow-providers-postgres>=5.10.0

